+++ src/cpp/utils/cpuset_to_cstr.cpp
@@ -9,10 +9,10 @@
 char *tbsla::cpp::utils::cpuset_to_cstr(cpu_set_t *mask, char *str)
 {
     char *ptr = str;
-    int i, j, entry_made = 0;
+    long long int i, j, entry_made = 0;
     for (i = 0; i < CPU_SETSIZE; i++) {
         if (CPU_ISSET(i, mask)) {
-            int run = 0;
+            long long int run = 0;
             entry_made = 1;
             for (j = i + 1; j < CPU_SETSIZE; j++) {
                 if (CPU_ISSET(j, mask)) run++;
